Enhance MarketDataProvider with ETH price resolution via WETH->QUOTE path
This commit updates the MarketDataProvider to include logic for resolving ETH prices using a WETH to QUOTE token path. It also clarifies the pricing logic in the documentation, specifying that DIEM is resolved via TRADE_PATH, VVV via QUOTE_TOKEN_ADDRESS, and USDC remains at a constant value of 1.0. These changes improve the accuracy and reliability of price retrieval for ETH, enhancing the overall functionality of market data interactions.

+++ apps/broker-api/tenant_store.py
@@ -13,7 +13,7 @@ class Tenant:
     id: str
     label: str
     subkey: str
-    quota: int
+    quota: float
     expires_at: Optional[str] = None
     status: str = "active"  # active|revoked
     # Optional Venice API key id to allow revoke via API
